move NotionConverter inside plugin
